25/05

+++ app/src/main/java/com/example/syscall/data/model/CallModel.kt
@@ -1,3 +1,3 @@
 package com.example.syscall.data.model
 
-data class CallModel (val idBell:Int, val callOption:Int)
+data class CallModel (val idBell:Long, val callOption:Int)
